fred_macro_intelligence/requirements.txt
pandas>=2.0
numpy>=1.24
fredapi>=0.5
lightgbm>=4.0
xgboost>=2.0
scikit-learn>=1.3
torch>=2.0
hmmlearn>=0.3
shap>=0.44
plotly>=5.18
dash>=2.14
dash-bootstrap-components>=1.5
kaleido>=0.2              # plotly static export (PNG/SVG)
scipy>=1.11
statsmodels>=0.14
requests>=2.31
pyyaml>=6.0
python-dotenv>=1.0
joblib>=1.3
ipywidgets>=8.0
jupyter>=1.0
